api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend"
authors = [
    {name = "dev"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "openai (>=2.26.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
